#234 feat: Implement chat functionality for user and restaurant communication
