+++ src/main/java/kr/oshino/eataku/reservation/user/controller/ReservationUserController.java
@@ -257,21 +257,15 @@ public class ReservationUserController {
         return "reservation/reservationDetail";
     }
 
-    // 쩝쩝채팅
+    // 유저 채팅
     @GetMapping("/user/chatting/{restaurantNo}")
     public String chattingView(@PathVariable String restaurantNo, Model model){
 
-//        String userType = "user";
-
-//        model.addAttribute("userType", userType);
-        model.addAttribute("restaurantNo", restaurantNo);
-
+        model.addAttribute("roomId", restaurantNo);
+        model.addAttribute("userType", "user");
         log.info("🍎restaurantNo = " + restaurantNo);
-//        log.info("\uD83C\uDF4EuserType = " + userType);
-
-        return "restaurant/chatting";
+        return "ws/user-chat";
     }
-
 
 }
