--- src/main/java/kr/oshino/eataku/restaurant/admin/controller/ChatController.java
@@ -1,22 +1,0 @@
-package kr.oshino.eataku.restaurant.admin.controller;
-
-import kr.oshino.eataku.restaurant.admin.model.dto.ChatMessageDTO;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@Slf4j
-@RequiredArgsConstructor
-public class ChatController {
-
-    private final SimpMessagingTemplate messagingTemplate;
-
-    @MessageMapping("/chat/message")
-    public void message(ChatMessageDTO message){
-        log.info("🍎Received message: {}", message);
-        messagingTemplate.convertAndSend("/sub/chat/room/" + message.getRoomId(), message);
-    }
-}
