+++ src/main/java/kr/oshino/eataku/restaurant/admin/controller/RestaurantAdminController.java
@@ -1,18 +1,15 @@
 package kr.oshino.eataku.restaurant.admin.controller;
 
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import kr.oshino.eataku.common.enums.FoodType;
 import kr.oshino.eataku.common.enums.HashTag;
 import kr.oshino.eataku.member.model.dto.CustomMemberDetails;
 import kr.oshino.eataku.reservation.user.model.dto.responseDto.*;
 import kr.oshino.eataku.reservation.user.service.ReservationUserService;
-import kr.oshino.eataku.restaurant.admin.entity.RestaurantInfo;
 import kr.oshino.eataku.restaurant.admin.model.dto.*;
 import kr.oshino.eataku.restaurant.admin.service.RestaurantAdminService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
@@ -22,7 +19,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.sql.Date;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.temporal.TemporalAdjusters;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -447,56 +443,6 @@ public class RestaurantAdminController {
         return "restaurant/preview";
     }
 
-    /***
-     * 매출 조회
-     * @param model
-     * @param startDd
-     * @param endDd
-     * @return
-     */
-//    @GetMapping("/main")
-//    public String getSalesReport(Model model,
-//                                 @RequestParam(value = "startDate", required = false) String startDd,
-//                                 @RequestParam(value = "endDate", required = false) String endDd) {
-//
-//        CustomMemberDetails member = (CustomMemberDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//        Long loginedRestaurantNo = member.getRestaurantNo();
-//
-//        LocalDateTime startDate;
-//        LocalDateTime endDate;
-//
-//        if(startDd == "" || startDd == null) {
-//            startDate = LocalDateTime.now().with(TemporalAdjusters.firstDayOfMonth());
-//        }else{
-//            startDate = LocalDateTime.parse(startDd);
-//        }
-//        if(endDd == "" || endDd == null) {
-//            endDate = LocalDateTime.now().with(TemporalAdjusters.lastDayOfMonth());
-//        }else {
-//            endDate = LocalDateTime.parse(endDd);
-//        }
-//
-//
-//        List<SalesDTO> sales = restaurantAdminService.selectSalesReport(startDate, endDate, loginedRestaurantNo);
-//
-//        model.addAttribute("loginedRestaurantNo", loginedRestaurantNo);
-//        model.addAttribute("sales", sales);
-//        model.addAttribute("startDate", startDate);
-//        model.addAttribute("endDate", endDate);
-//        log.info("\uD83C\uDF4E controller model: {}", model);
-//
-//        log.info("\uD83C\uDF4E startDate: {}, endDate: {}, restaurantNo: {}", startDate, endDate, loginedRestaurantNo);
-//
-//        return "restaurant/main";
-//    }
 
-    /***
-     * 채팅
-     * @return
-     */
-    @GetMapping("/chatting")
-    public String chattingView(){
-        return "restaurant/chatting";
-    }
 
 }
