+++ src/main/java/kr/oshino/eataku/ws/controller/ChatController.java
@@ -1,0 +1,44 @@
+package kr.oshino.eataku.ws.controller;
+
+import kr.oshino.eataku.member.model.dto.CustomMemberDetails;
+import kr.oshino.eataku.ws.dto.ChatMessageDTO;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Controller
+@Slf4j
+@RequiredArgsConstructor
+public class ChatController {
+
+    private final SimpMessagingTemplate messagingTemplate;
+
+    @MessageMapping("/chat/message")
+    public void message(ChatMessageDTO message){
+        log.info("🍎Received message: {}", message);
+        messagingTemplate.convertAndSend("/sub/chat/room/" + message.getRoomId(), message);
+    }
+
+    /***
+     * 식당 채팅
+     * @return
+     */
+    @GetMapping("restaurant/chatting")
+    public String chattingView(Model model){
+
+        CustomMemberDetails member = (CustomMemberDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Long restaurantNo = member.getRestaurantNo();
+
+        model.addAttribute("userType", "restaurant");
+        model.addAttribute("roomId", restaurantNo);
+        log.info("🍎userType = " + model);
+        return "ws/restaurant-chat";
+    }
+
+}
