+++ src/main/resources/templates/restaurant/main.html
@@ -75,26 +75,6 @@
                                             <td th:text="${sale.time}">18:00</td>
                                             <td th:text="${sale.serviceType}">웨이팅</td>
                                         </tr>
-<!--                                        <tr>-->
-<!--                                            <td>2024-06-01</td>-->
-<!--                                            <td>2</td>-->
-<!--                                            <td>18:00</td>-->
-<!--                                            <td>예약</td>-->
-<!--                                        </tr>-->
-                                        <!-- Add more rows as needed -->
-
-<!--                                     <tbody id="salesTableBody">
-                                        <tr th:each="sale:${sales}" class="tableData">
-                                            <input type="hidden" th:value="${sale.customerName}">
-                                            <input type="hidden" th:value="${sale.customerContact}">
-                                            <input type="hidden" th:value="${sale.usageType}">
-                                            <input type="hidden" th:value="${sale.numberOfPeople}">
-                                            <input type="hidden" th:value="${sale.usageCount}">
-                                            <td th:text="${sale.salesDate}"></td>
-                                            <td th:text="${sale.numberOfPeople}"></td>
-                                            <td th:text="${sale.usageTime}"></td>
-                                            <td th:text="${sale.usageType}"></td>
-                                        </tr> -->
 
                                     </tbody>
                                 </table>
