+++ src/main/resources/templates/ws/restaurant-chat.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <meta charset="UTF-8">
-        <title>채팅</title>
+        <title>식당 채팅</title>
         <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
         <link rel="stylesheet" href="/css/fragment/navbar.css">
         <link rel="stylesheet" href="/css/fragment/footer.css">
@@ -17,25 +17,18 @@
                 <img src="/images/logo.png" alt="icon">
                 <input type="text" id="messageInput" placeholder="메세지를 입력해주세요." maxlength="1000" oninput="updateCharCount()">
                 <span class="char-count" id="charCount">0/1000</span>
-                <button onclick="sendMessage()">전송</button>
+                <button th:onclick="sendMessage([[${roomId}]])">전송</button>
+                <input type="hidden" id="roomId" th:data-roomId="${roomId}">
+                <input type="hidden" id="sender" th:data-sender="${userType}">
             </div>
         </div>
-        <script>
-            var userType = /*[[${userType}]]*/ 'default'; // 서버에서 유저 타입 전달
-
-            if (userType === 'user') {
-                $.getScript('/js/restaurant/user-chat.js');
-            } else if (userType === 'restaurant') {
-                $.getScript('/js/restaurant/restaurant-chat.js');
-            }
-        </script>
 
         <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
         <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/sockjs-client@1.5.2/dist/sockjs.min.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js"></script>
-        <script src="/js/restaurant/user-chat.js"></script>
+        <script src="/js/ws/restaurant-chat.js"></script>
 
     </body>
