+++ src/main/resources/templates/ws/user-chat.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <title>유저 채팅</title>
+        <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
+        <link rel="stylesheet" href="/css/fragment/navbar.css">
+        <link rel="stylesheet" href="/css/fragment/footer.css">
+        <link rel="stylesheet" href="/css/restaurant/chatting.css">
+    </head>
+    <body>
+        <div class="chat-container">
+            <div class="chat-messages" id="chatMessages">
+                <!-- 여기에 채팅 메시지가 표시됩니다 -->
+            </div>
+            <div class="chat-input">
+                <img src="/images/logo.png" alt="icon">
+                <input type="text" id="messageInput" placeholder="메세지를 입력해주세요." maxlength="1000" oninput="updateCharCount()">
+                <span class="char-count" id="charCount">0/1000</span>
+                <button onclick="sendMessage()">전송</button>
+            </div>
+        </div>
+
+        <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
+        <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+        <script src="https://cdn.jsdelivr.net/npm/sockjs-client@1.5.2/dist/sockjs.min.js"></script>
+        <script src="https://cdn.jsdelivr.net/npm/stompjs@2.3.3/lib/stomp.min.js"></script>
+        <script src="/js/ws/user-chat.js"></script>
+
+    </body>
+
+</html>
